+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Todo Tree Change Log
 
 - Sort tags only view on tags order by default
+- Ignore case when validating HTML colour names
 
 ## v0.0.205 - 2021-03-13